// hub-api/models/api_editionDiagramsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { api_editionDiagramVolume } from './api_editionDiagramVolume';
export type api_editionDiagramsResponse = {
    hasNoDiagrams?: boolean;
    imageUrlsByName?: Record<string, string>;
    key?: string;
    volumes?: Array<api_editionDiagramVolume>;
};